+++ tests/parse/array_literal.js
@@ -1,15 +1,15 @@
 var lexer = require('../../index').lex.lexer;
 var parser = require('../../index').parse.parser;
 
-
 var testParser = function(stream) {
-    var result = parser.parseStream(stream);
+    var result = parser.parseStream(lexer.lex(stream));
     return result.body[0].expression;
 };
 
 
 exports.empty = function(test) {
-    var result = testParser(lexer.lex("[];"));
+    var result = testParser("[];");
+    
     test.equal(result.elements.length, 0);
     
     test.done();
@@ -17,7 +17,8 @@ exports.empty = function(test) {
 
 
 exports.single_element = function(test) {
-    var result = testParser(lexer.lex("[3];"));
+    var result = testParser("[3];");
+    
     test.equal(result.elements.length, 1);
     test.equal(result.elements[0].value, 3);
     
@@ -25,7 +26,8 @@ exports.single_element = function(test) {
 };
 
 exports.multi_element = function(test) {
-    var result = testParser(lexer.lex("[3, 4];"));
+    var result = testParser("[3, 4];");
+    
     test.equal(result.elements.length, 2);
     test.equal(result.elements[0].value, 3);
     test.equal(result.elements[1].value, 4);